[wallet-refactor-3] Move AccountDetails to new module #2058 (#2059)
* Move utils

* Split account in smaller components

* Split account in smaller components

* Fix code style issues with Prettier

---------
