+++ src/cow-react/modules/account/containers/Transaction/styled.ts
@@ -1,7 +1,6 @@
 import styled, { css, keyframes } from 'styled-components/macro'
 import { StyledSVG } from 'components/Loader'
-import { LinkStyledButton, StyledLink } from 'theme'
-import { TransactionState as OldTransactionState } from '../AccountDetails/TransactionMod'
+import { ExternalLink, LinkStyledButton, StyledLink } from 'theme'
 import { RowFixed } from 'components/Row'
 import { transparentize } from 'polished'
 import { StyledLogo } from 'components/CurrencyLogo'
@@ -302,7 +301,7 @@ export const StatusLabel = styled.div<{
     opacity: 0.15;
   }
 
-  ${({ theme, isCancelling, isPresignaturePending, isTransaction, isPending, isCreating }) =>
+  ${({ theme, isCancelling, isPresignaturePending, isTransaction, isPending }) =>
     (isCancelling || isPresignaturePending || (isPending && isTransaction)) &&
     css`
       &::after {
@@ -349,6 +348,18 @@ export const StatusLabelBelow = styled.div<{ isCancelling?: boolean }>`
     opacity: 1;
     color: ${({ theme }) => theme.text1};
   }
+`
+
+export const OldTransactionState = styled(ExternalLink)<{ pending: boolean; success?: boolean }>`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  /* text-decoration: none !important; */
+  border-radius: 0.5rem;
+  padding: 0.25rem 0rem;
+  font-weight: 500;
+  font-size: 0.825rem;
+  color: ${({ theme }) => theme.primary1};
 `
 
 // override the href, pending and success props
