+++ src/cow-react/modules/swap/pure/EthFlow/EthFlowModalContent/EthFlowModalBottomContent.tsx
@@ -1,6 +1,6 @@
 import { WrappingPreview, WrappingPreviewProps } from '../WrappingPreview'
 
-import SimpleAccountDetails from '@cow/modules/account/containers/AccountDetails/SimpleAccountDetails'
+import SimpleAccountDetails from '@cow/modules/account/containers/SimpleAccountDetails'
 
 import { ActionButton } from './ActionButton'
 import { EthFlowState } from '../../../services/ethFlow/types'
