+++ src/cow-react/modules/wallet/api/components/WalletModal/options/WalletConnectOption.tsx
@@ -1,7 +1,7 @@
 import { Connector } from '@web3-react/types'
 import WALLET_CONNECT_ICON_URL from '../../../assets/walletConnectIcon.svg'
 import { ConnectionType, walletConnectConnection } from 'connection'
-import { getConnectionName } from 'connection/utils'
+import { getConnectionName } from '@cow/modules/wallet/api/utils'
 import { useIsActiveWallet } from 'hooks/useIsActiveWallet'
 
 import Option from '../Option'
