+++ src/cow-react/modules/wallet/api/utils/connections.ts
@@ -10,7 +10,7 @@ import { SupportedChainId } from 'constants/chains'
 import Fortmatic from 'fortmatic'
 
 import UNISWAP_LOGO_URL from '../assets/svg/logo.svg'
-import { RPC_URLS } from '../constants/networks'
+import { RPC_URLS } from 'constants/networks'
 
 export enum ConnectionType {
   INJECTED = 'INJECTED',
