+++ src/custom/components/TransactionConfirmationModal/TransactionConfirmationModalMod.tsx
@@ -15,7 +15,7 @@ import { isL2ChainId } from 'utils/chains'
 import Circle from 'assets/images/blue-loader.svg'
 import { CloseIcon, CustomLightSpinner, ExternalLink } from 'theme'
 // import { ExplorerDataType, getExplorerLink } from 'utils/getExplorerLink'
-import { TransactionSummary } from '@cow/modules/account/containers/AccountDetails/TransactionSummary'
+import { TransactionSummary } from '@cow/modules/account/containers/TransactionSummary'
 import { /* ButtonLight */ ButtonPrimary } from '../Button'
 import { AutoColumn, ColumnCenter } from 'components/Column'
 // import Modal from 'components/Modal'
